refactor: folder 구조 project convention에 맞춰 변경
