+++ src/main/java/io/backend/blogproject/controller/CategoryController.java
@@ -1,0 +1,4 @@
+package io.backend.blogproject.controller;
+
+public class CategoryController {
+}
