+++ src/main/java/io/backend/blogproject/repository/CategoryRepository.java
@@ -1,0 +1,4 @@
+package io.backend.blogproject.repository;
+
+public class CategoryRepository {
+}
